requirements.txt

numpy<2.0
mediapipe==0.10.11
opencv-python
pyobjc
pyobjc-framework-Quartz
pyobjc-framework-AppKit
pyobjc-framework-ApplicationServices
